requirements.txt
absl-py==0.10.0
addict==2.2.1
albumentations==0.4.5
arrow==0.15.5
boto3==1.12.39
botocore==1.15.39
cachetools==4.1.1
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.1
cloudpickle==1.3.0
cycler==0.10.0
Cython==0.29.17
cytoolz==0.10.1
dask==2.13.0
decorator==4.4.2
decord==0.3.9
dicom2nifti==2.2.10
docutils==0.15.2
dominate==2.5.1
fastremap==1.10.2
ffmpeg==1.4
ffprobe==0.5
filelock==3.0.12
fill-voids==2.0.0
future==0.18.2
gdown==3.12.2
google-auth==1.20.1
google-auth-oauthlib==0.4.1
grpcio==1.31.0
idna==2.9
image-classifiers==0.2.2
imagecodecs==2020.2.18
imageio==2.8.0
imageio-ffmpeg==0.4.2
imgaug==0.2.6
importlib-metadata==1.7.0
iterative-stratification==0.1.6
jmespath==0.9.5
joblib==0.14.1
jpegio==0.2.1
jsonpatch==1.25
jsonpointer==2.0
kaggle==1.5.6
Keras==2.4.3
kiwisolver==1.1.0
llvmlite==0.32.0
-e git+https://github.com/JoHof/lungmask.git@c54c28c7e6e1fe224d41ca374623b39a667c3137#egg=lungmask
lxml==4.5.0
Markdown==3.2.2
matplotlib==3.1.3
mdai==0.5.0
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
mmcv==1.0.4
mmcv-full==1.0.4
# Editable install with no version control (mmdet==1.1.0+1b299b5)
-e /home/ianpan/anaconda3/envs/pytorch_p37_nt/lib/python3.7/site-packages/mmdet-1.1.0+1b299b5-py3.7.egg
munch==2.5.0
networkx==2.4
nibabel==3.1.0
numba==0.49.0
numpy==1.18.4
oauthlib==3.1.0
olefile==0.46
omegaconf==2.0.0
opencv-python==4.2.0.34
opencv-python-headless==4.2.0.34
openslide-python==1.1.1
packaging==20.3
pandas==1.0.3
Pillow==7.0.0
PIMS==0.4+209.g19ad451
pretrainedmodels==0.7.4
protobuf==3.13.0
pyarrow==0.16.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.0
pycparser==2.20
pydicom==1.4.2
PyExecJS==1.5.1
pyparsing==2.4.6
PySocks==1.7.1
python-dateutil==2.8.1
python-slugify==4.0.0
pytorch-lightning==0.9.1.dev0
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==19.0.1
regex==2020.4.4
requests==2.23.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.6
s3transfer==0.3.3
sacremoses==0.0.38
scikit-image==0.16.2
scikit-learn==0.22.1
scikit-video==1.1.11
scipy==1.4.1
sentencepiece==0.1.85
SimpleITK==1.2.4
six==1.14.0
slicerator==1.0.0
tensorboard==2.2.0
tensorboard-plugin-wit==1.7.0
terminaltables==3.1.0
text-unidecode==1.3
tifffile==2020.2.16
timm==0.2.1
tokenizers==0.9.2
toolz==0.10.0
torch==1.6.0.dev20200506
torchfile==0.1.0
torchvision==0.7.0.dev20200506
tornado==6.0.4
tqdm==4.44.1
transformers==3.4.0
translators==4.4.3
typeguard==2.9.1
typing-extensions==3.7.4.3
urllib3==1.24.3
visdom==0.1.8.9
vtk==9.0.1
websocket-client==0.57.0
Werkzeug==1.0.1
yapf==0.30.0
zipp==3.1.0
